Creación frontend Angular

+++ mochicloset-backend/Datos/Publicacion.cs
@@ -5,7 +5,7 @@ public class Publicacion
     public int Id { get; set; }
     public string? Titulo { get; set; }
     public string? Descripcion { get; set; }
-    public float Precio { get; set; }
+    public decimal Precio { get; set; }
     public string? Talla { get; set; }
     public string? Condicion { get; set; }
     public string? Estado { get; set; }
